Forum: Déplacement d'une methode de C vers M
